requirements-base.txt
alembic==1.14.0
babel==2.16.0
backports.tarfile==1.2.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.18
docutils==0.21.2
Flask==3.1.0
flask-babel==4.0.0
Flask-Limiter==3.10.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
flask-session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Table==0.5.0
fonttools==4.55.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
isort==7.0.0
pylint==4.0.2
ruff==0.14.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
kiwisolver==1.4.7
limits==4.0.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.20
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pkginfo==1.12.0
plotly==5.24.1
psycopg2-binary==2.9.10
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
readme_renderer==44.0
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.36
tenacity==8.2.2
twine==6.0.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0